// types/schema.ts
export interface components {
  schemas: {
// ...
    Provider: {
      /** @description Id of the provider */
      id?: string;
      descriptor?: components['schemas']['Descriptor'];
      /** @description Category Id of the provider */
      category_id?: string;
      rating?: components['schemas']['Rating']['value'];
      time?: components['schemas']['Time'];
      categories?: components['schemas']['Category'][];
      fulfillments?: components['schemas']['Fulfillment'][];
      payments?: components['schemas']['Payment'][];
      locations?: (components['schemas']['Location'] & {
        rateable?: components['schemas']['Rateable'];
      })[];
      offers?: components['schemas']['Offer'][];
      items?: components['schemas']['Item'][];
      /**
       * Format: date-time
       * @description Time after which catalog has to be refreshed
       */
      exp?: string;
      rateable?: components['schemas']['Rateable'];
      tags?: components['schemas']['Tags'];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}